[update] http / 分かりやすさのため,旧Http classの関数を1ファイルにまとめた
